tests/wpt/web-platform-tests/keyboard-lock/navigator-keyboardLock.https.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
'use strict';

promise_test(() => {
  const p = navigator.keyboardLock(['a', 'b']);
  assert_true(p instanceof Promise);
  return p;
}, 'Keyboard Lock keyboardLock');

</script>
